+++ packages/brick_graphql/lib/graphql.dart
@@ -2,6 +2,7 @@ export 'package:brick_graphql/src/graphql_adapter.dart';
 export 'package:brick_graphql/src/graphql_model.dart';
 export 'package:brick_graphql/src/graphql_model_dictionary.dart';
 export 'package:brick_graphql/src/graphql_provider.dart';
+export 'package:brick_graphql/src/runtime_graphql_definition.dart';
 
 // Annotations
 export 'package:brick_graphql/src/annotations/graphql_serializable.dart';