+++ packages/brick_offline_first_with_graphql_build/lib/builder.dart
@@ -30,6 +30,8 @@ Builder offlineFirstAggregateBuilder(options) => AggregateBuilder(requiredImport
       "import 'package:brick_sqlite_abstract/db.dart';",
       "import 'package:brick_core/query.dart';",
       "import 'package:gql/language.dart' as lang;",
+      "import 'package:brick_graphql/graphql.dart' show RuntimeGraphqlDefinition;",
+      "import 'package:brick_offline_first_with_graphql/offline_first_with_graphql.dart' show OfflineFirstWithGraphqlRepository, OfflineFirstWithGraphqlAdapter;",
     ]);
 Builder offlineFirstAdaptersBuilder(options) =>
     AdapterBuilder<ConnectOfflineFirstWithGraphql>(offlineFirstGenerator);
@@ -38,9 +40,9 @@ Builder offlineFirstModelDictionaryBuilder(options) =>
       const OfflineFirstModelDictionaryGenerator('Graphql'),
       expectedImportRemovals: [
         "import 'package:brick_offline_first_abstract/annotations.dart';",
-        'import "package:brick_offline_first_abstract/annotations.dart";',
+        "import 'package:brick_offline_first_abstract/annotations.dart';",
         "import 'package:brick_offline_first/offline_first.dart';",
-        'import "package:brick_offline_first/offline_first.dart";',
+        "import 'package:brick_offline_first/offline_first.dart';",
       ],
     );
 Builder offlineFirstNewMigrationBuilder(options) => OfflineFirstMigrationBuilder();